requirements.txt
redis~=4.5.5
kafka-python==2.0.2
boto3~=1.26.133
requests~=2.30.0